requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
distlib==0.3.4
filelock==3.6.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.1.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.15
urllib3==1.26.9
virtualenv==20.14.0
Werkzeug==2.0.1
